src/client/requirements.txt
django==3.0
biopython==1.76
python-memcached==1.59
mmh3==2.5.1
bitarray==1.2.0
requests==2.22.0